+++ PolygonChecker/triangleSolver.h
@@ -1,2 +1,2 @@
 #pragma once
-char* analyzeTriangle(int side1, int side2, int side3);
+char* analyzeTriangle(double side1, double side2, double side3);
